backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==22.0.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
boto3==1.34.100
httpx==0.27.0
slowapi==0.1.9
cryptography==42.0.5
groq==0.9.0
transformers==4.41.0
torch==2.3.0
spacy==3.7.4
pydub==0.25.1
websockets==12.0
psycopg2-binary==2.9.9
pytest==8.2.0
pytest-asyncio==0.23.6
moto[s3,dynamodb,transcribe]==5.0.7
